Move pkg/kubectl/util/qos to staging
