+++ pkg/kubectl/describe/versioned/describe.go
@@ -67,13 +67,13 @@ import (
 	"k8s.io/klog"
 	deploymentutil "k8s.io/kubectl/pkg/util/deployment"
 	"k8s.io/kubectl/pkg/util/event"
+	"k8s.io/kubectl/pkg/util/qos"
 	"k8s.io/kubectl/pkg/util/rbac"
 	storageutil "k8s.io/kubectl/pkg/util/storage"
 	"k8s.io/kubernetes/pkg/kubectl/describe"
 	"k8s.io/kubernetes/pkg/kubectl/scheme"
 	"k8s.io/kubernetes/pkg/kubectl/util/certificate"
 	"k8s.io/kubernetes/pkg/kubectl/util/fieldpath"
-	"k8s.io/kubernetes/pkg/kubectl/util/qos"
 	resourcehelper "k8s.io/kubernetes/pkg/kubectl/util/resource"
 	"k8s.io/kubernetes/pkg/kubectl/util/slice"
 )
